Create crowdstrike-fix.sh
This script will output the logs to the Desktop for any local user to verify the removal of the files. I created a system service to auto run this as soon as the bootable USB loads it runs the script and set up ssh access in the event the script fails.

IM VERY AWARE MICROSOFT RELEASED A TOOL.

This script was primarily made to act a as a back up in the event that recovery USB fails on some systems.